feature: criando alguns testes para os calculators
