+++ TheatricalPlayersRefactoringKata.Tests/IntegrationTests/StatementPrinterTests.TestStatementExampleLegacy.approved.txt
