+++ TheatricalPlayersRefactoringKata.Tests/IntegrationTests/StatementPrinterTests.TestTextStatementExample.approved.txt
