+++ TheatricalPlayersRefactoringKata.Tests/IntegrationTests/StatementPrinterTests.TestXmlStatementExample.approved.txt
